Connect all views to patient dashboard (#24)
# *Connect all views to patient dashboard*

## :recycle: Current situation & Problem
This is a follow-up PR to: 
1. [Create dashboard view with progress
rings](https://github.com/CS342/2025-Stanford-360/pull/21)
2. [Turn `PatientManager` into Spezi
Module](https://github.com/CS342/2025-Stanford-360/pull/22)


Currently, the DashboardView's data is static since the patient model is
not yet connected to the rest of the app. This means that updates to the
patient's activity, hydration, or protein in their respective tabs do
not reflect in the dashboard.


## :gear: Release Notes 
* Uses the `PatientManager` Module to retrieve the day's activity
minutes, hydration ounces, and protein grams from the respective models'
views
* Updates the DashboardView's data accordingly
<img width="377" alt="Screenshot 2025-02-28 at 12 56 09 PM"
src="https://github.com/user-attachments/assets/80995d67-32d3-4634-bd45-8cb07cf13e14"
/>

In more follow-up PRs, we will:
* Fetch the data from Firestore on initial load of the Dashboard View,
such that it does not rely on the other views to fetch data first
* Continue to iterate on and improve the UI

## :books: Documentation
N/A


## :white_check_mark: Testing
Testing was done locally. Written tests are on pause until we understand
what is causing them to pass locally but break in the CI.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md):
- [X] I agree to follow the [Code of
Conduct](https://github.com/CS342/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/CS342/.github/blob/main/CONTRIBUTING.md).

+++ Stanford360/Dashboard/Models/Patient.swift
@@ -14,5 +14,5 @@ import Foundation
 struct Patient {
 	var activityMinutes: Int
 	var hydrationOunces: Double
-	var proteinGrams: Int
+	var proteinGrams: Double
 }
